requirements.txt
# STREAMLINED PRODUCTION REQUIREMENTS - Memory-optimized for Render deployment
# Core FastAPI Application
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Configuration & Environment
python-dotenv==1.1.1
pydantic==2.11.7
pydantic-settings==2.10.1

# Database & ORM - Essential for all endpoints
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
asyncpg==0.30.0
pgvector==0.2.5

# Authentication & Security
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
fastapi-users[sqlalchemy]==13.0.0
email-validator==2.1.0
httpx-oauth==0.14.1
cryptography==42.0.8
pyotp==2.9.0
qrcode[pil]==7.4.2

# HTTP & Communication
httpx==0.27.2
requests==2.32.3
requests-oauthlib==1.3.1
aiohttp==3.10.11
sse-starlette==2.1.3
websockets==12.0

# Logging & Monitoring
structlog==25.4.0
sentry-sdk[fastapi]==2.35.0
prometheus-client==0.20.0

# Essential AI - OpenAI only for core functionality
openai==1.75.0

# Social Media APIs
tweepy==4.14.0

# Task Queue - Essential for background processing
celery==5.4.0
redis==5.2.1

# File Processing & Utilities
beautifulsoup4==4.12.3
Pillow==11.0.0
python-magic==0.4.27

# Development & Testing (commented out for production builds)
# pytest==8.3.3
# pytest-cov==4.0.0
# pytest-asyncio==0.21.1
# pytest-mock==3.12.0
# pytest-env==1.1.5
# pytest-timeout==2.3.1
# factory-boy==3.3.1
# faker==25.9.2

# Heavy AI Libraries (temporarily disabled to resolve build failures)
# crewai==0.126.0
# crewai-tools==0.47.0
# numpy==1.26.4
# pytrends==4.9.2
# flower==2.0.1
# aiosqlite==0.19.0
# prometheus-fastapi-instrumentator==7.0.0